service test update

+++ src/test/java/es/cic/curso25/proyecto009/service/ArbolServiceIntegrationTest.java
@@ -98,9 +98,10 @@ public class ArbolServiceIntegrationTest {
     @BeforeEach
     private void deleteIfExist() {
         List<Arbol> arboles = arbolService.getByAll();
+        
 
         arboles.forEach(arbol -> {
-            if (arbol.getNombre().equals("Espeletia"))
+            if (arbol!=null&& arbol.getNombre()!=null && arbol.getNombre().equals("Espeletia"))
 
                 arbolService.deleteById(arbol.getId());
         });
